swing_trader/requirements.txt
numpy>=1.26
pandas>=2.3
pyarrow>=15.0
matplotlib>=3.10
ib_async>=2.1
asyncpg>=0.31
pydantic>=2.12
pyyaml>=6.0
requests>=2.31
